requirements.txt
ezdxf>=1.4.1
matplotlib>=3.8.1
pillow>=10.0.1